+++ CHANGES.md
@@ -9,7 +9,7 @@ SPDX-License-Identifier: MIT OR Apache-2.0
 
 * Fix IPv4 TTL
 
-* Use String#compareTo instead of String#equals for creating Service.
+* Remove incubating annotation on restricted methods.
 
 
 ## Version 1.3.1 (2021-05-22)